Refactor/#17:도메인 재설계 1. JoinType 컬럼을 클럽 Entity 로 이동 2. 기존에 사용했던 ApplicationForm Entity 삭제
