--- cluvr-api/src/main/java/com/example/cluvrapi/domain/applicationForm/repository/ApplicationFormRepository.java
@@ -1,8 +1,0 @@
-package com.example.cluvrapi.domain.applicationForm.repository;
-
-import com.example.cluvrapi.domain.applicationForm.entity.ApplicationForm;
-import com.example.cluvrapi.domain.common.repository.BaseRepository;
-
-public interface ApplicationFormRepository extends BaseRepository<ApplicationForm, Long> {
-
-}
