+++ cluvr-api/src/main/java/com/example/cluvrapi/domain/club/enums/JoinType.java
@@ -1,0 +1,8 @@
+package com.example.cluvrapi.domain.club.enums;
+
+public enum JoinType {
+	PROBLEM_FORM, //문제 양식 제출
+	SUBMISSION_FORM,   //가입 신청서 제출
+	DIRECT_JOIN,        //가입 신청과 동시에 바로 가입
+	SIMPLE_REQUEST      //단순 가입 신청
+}
