--- cluvr-api/src/main/java/com/example/cluvrapi/domain/joinRequest/enums/JoinType.java
@@ -1,8 +1,0 @@
-package com.example.cluvrapi.domain.joinRequest.enums;
-
-public enum JoinType {
-    PROBLEM_SUBMISSION, //문제 양식 제출
-    APPLICATION_FORM,   //가입 신청서 제출
-    DIRECT_JOIN,        //가입 신청과 동시에 바로 가입
-    SIMPLE_REQUEST      //단순 가입 신청
-}
